Reworded missing values based on Tomer's PR
For simplicity we abandoned https://github.com/nheeren/material_intensity_db/pull/3
 and added the changes here.

New Figure in codebook.md to illustrate the decision tree for missing values.

Reworded and renamed the Special values section.

Added an introduction to the Special values section.

Disclaimer in README for v1.0 release, which will be archived for the publication.